updated makeRelease.py script to include boost dlls

+++ makeRelease.py
@@ -2,6 +2,7 @@ from pathlib import Path, PurePath
 import os
 import shutil
 import zipfile
+import sys
 
 PROJECT_NAME = "rrt"
 BIN_DIR = "release"
@@ -11,13 +12,21 @@ RELEASE_DIR = BIN_DIR + "/release"
 CMAKE_PRE_BUILD_CMD = "cmake . -B" + BIN_DIR
 CMAKE_BUILD_CMD = "cmake --build " + BIN_DIR + " --config MinSizeRel -j4"
 
+BOOST_VERSION = "1_71"
+BOOST_PREFIX = "libboost_"
+BOOST_SUFFIX = "-mgw73-mt-x64-" + BOOST_VERSION + ".dll"
+
+BOOST_LIBS = ["date_time", "filesystem",
+              "thread", "log"]  # change it to include in zip
+
 IGNORED_FILES = ["D3Dcompiler_47.dll",
-                 "opengl32sw.dll", "libEGL.dll", "libGLESV2.dll"]
+                 "opengl32sw.dll", "libEGL.dll", "libGLESV2.dll"]  # these files will be deleted from release
 
 
-def cmake():
-    if Path(BIN_DIR).exists():
-        shutil.rmtree(BIN_DIR)
+def cmake(clean=False):
+    if clean:
+        if Path(BIN_DIR).exists():
+            shutil.rmtree(BIN_DIR)
     os.system(CMAKE_PRE_BUILD_CMD)
     os.system(CMAKE_BUILD_CMD)
 
@@ -44,6 +53,9 @@ def addToZip(version=""):
                 if filename not in IGNORED_FILES:
                     releaseZip.write(dirpath + "/" + filename,
                                      fullPathWithoutReleaseDir)
+        for lib in BOOST_LIBS:
+            fullName = BOOST_PREFIX + lib + BOOST_SUFFIX
+            releaseZip.write(findFile("c:/boost/lib", fullName), fullName)
     print('Zip created in "' + zipPath + '"')
 
 
@@ -68,7 +80,11 @@ def getProjectVersion():
 
 
 if __name__ == "__main__":
-    cmake()
+    if (len(sys.argv) == 0):
+        cmake(clean=False)
+    elif (sys.argv[0] == "--clean"):
+        cmake(clean=True)
+
     prepareReleaseDir()
     os.system(findFile(os.getenv("QT_DIR", "c:/Qt"),
                        "windeployqt.exe") + " --release " + RELEASE_DIR)
